+++ srcs/Commands/Privmsg.cpp
@@ -103,7 +103,7 @@ void HDE::SocketHde::sendMessageToAllForPrivmsg(int i, std::string channelname, 
             }
             std::string nick = clt.at(fds[i].fd).getNickname();
             std::string selfStr = ":" + nick  + "!" + nick + "@" + clt.at(fds[i].fd).getLocalhost() + " PRIVMSG " +  channelname + " :" + message + "\r\n";
-            for(int index = 0; index < add.size() ; index++)
+            for(size_t index = 0; index < add.size() ; index++)
                 sendMessage(selfStr, add.at(index));
         }
     }
